Transplant proven offsets back from feature/re
Bring back the small set of useful layout facts that were dropped when master was rebuilt from the 0.5.1 baseline.

Transplanted from the RE branch:
- GameStateBuffer.TOTAL_STATES = 13 to match the recovered fixed state-table shape
- VitalStruct stat-id and owner offsets at +0x08, +0x0C, +0x10, +0x20, +0x24, and +0x28
- InGameState.MsElapsed corrected to a 32-bit field at +0x400

Deliberately not transplanted:
- the larger runtime pattern / FF generation scaffolding
- broader naming and provenance-only annotations that do not affect the shipping reader

Validation:
- dotnet build D:\Work\Poe2 Helper.transplant\CheatCartridge.sln -p:TreatWarningsAsErrors=false -p:RestoreTreatWarningsAsErrors=false -p:WarningsAsErrors= -p:NoWarn=NU1605

+++ CheatCartridge/GameHelper/GameOffsets/States/InGameStateOffset.cs
@@ -14,10 +14,10 @@ public struct InGameStateOffset
     [FieldOffset(0x290)] public IntPtr AreaInstanceData;
     
     /// <summary>
-    /// Increasing number, probably count of ms elapsed since computer/client started
-    /// 0x328 => 0x400 (+D8)
+    /// Increasing integer timer maintained by the InGameState tick/update path.
+    /// Static RE and live reads both showed this is a 32-bit value.
     /// </summary>
-    [FieldOffset(0x400)] public IntPtr MsElapsed;
+    [FieldOffset(0x400)] public int MsElapsed;
     
     /// <summary>
     /// Contains ptr to Unicode string containing login server host.
